back/NewsCrawling/requirements.txt
beautifulsoup4
requests
pandas
python-dotenv
pymongo
fastapi[all]
uvicorn
selenium